refactor:年度判定スキームをclientと同様に
